feat: jwt open
